requirements.txt
astroid==2.4.2
gitdb==4.0.5
GitPython==3.1.11
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
munch==2.5.0
pylint==2.6.0
PyYAML==5.3.1
rope==0.18.0
six==1.15.0
smmap==3.0.4
toml==0.10.2
wrapt==1.12.1
